refactored the Maps and ErrandList Activities to extend the SingleFragmentActivity
