+++ app/src/main/java/com/amycohen/lab39masterdetailflow/ErrandListActivity.java
@@ -1,21 +1,11 @@
 package com.amycohen.lab39masterdetailflow;
 
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 
-public class ErrandListActivity extends AppCompatActivity {
+public class ErrandListActivity extends SingleFragmentActivity {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_single_fragment);
-
-        Fragment fragment = new ErrandListFragment();
-        FragmentManager fm = getSupportFragmentManager();
-        fm.beginTransaction()
-                .add(R.id.single_fragment_container, fragment)
-                .commit();
+    Fragment getFragment() {
+        return new ErrandListFragment();
     }
 }
