+++ app/src/main/java/com/amycohen/lab39masterdetailflow/MapsActivity.java
@@ -1,22 +1,11 @@
 package com.amycohen.lab39masterdetailflow;
 
 import android.support.v4.app.Fragment;
-import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
 
-public class MapsActivity extends AppCompatActivity  {
+public class MapsActivity extends SingleFragmentActivity  {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_single_fragment);
-
-        Fragment fragment = new MapsFragment();
-        FragmentManager fm = getSupportFragmentManager();
-        fm.beginTransaction()
-                .add(R.id.single_fragment_container, fragment)
-                .commit();
+    Fragment getFragment() {
+        return new MapsFragment();
     }
-
 }
