+++ app/src/main/java/com/amycohen/lab39masterdetailflow/SingleFragmentActivity.java
@@ -1,0 +1,24 @@
+package com.amycohen.lab39masterdetailflow;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v7.app.AppCompatActivity;
+
+public abstract class SingleFragmentActivity extends AppCompatActivity {
+
+    abstract Fragment getFragment();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_single_fragment);
+
+        Fragment fragment = getFragment();
+        FragmentManager fm = getSupportFragmentManager();
+        fm.beginTransaction()
+                .add(R.id.single_fragment_container, fragment)
+                .commit();
+    }
+
+}
